[Refactoring ⚙️] List에는 배열 순회만 하도록 분리
