+++ src/pages/ProjectDetailPage/components/Comments/CommentsItem.tsx
@@ -3,6 +3,7 @@
 import { useEffect, useState } from "react"
 import { SubmitHandler, useForm } from "react-hook-form"
 import { FieldErrors } from "react-hook-form"
+import { useNavigate } from "react-router-dom"
 
 import { Box, Flex, Text } from "@chakra-ui/react"
 import { Comment } from "api-models"
@@ -12,19 +13,26 @@ import { useEditCommentMutation } from "@pages/ProjectDetailPage/hooks/mutations
 import dateToTimeago from "@pages/ProjectDetailPage/utils/datetoTimeago"
 
 import { CommentFormValues } from "../../types/commentFormValues"
+import CommentsAvatar from "./CommentsAvatar"
 import CommentsButton from "./CommentsButton"
 import CommentsInputOrText from "./CommentsInputOrText"
 import { ProjectIdProps, withProjectId } from "./Hoc/withProjectId"
 
-interface CommentsContentProps extends ProjectIdProps {
+interface CommentsItemProps extends ProjectIdProps {
   comment: Comment
 }
 
 // eslint-disable-next-line react-refresh/only-export-components
-const CommentsContent = ({ comment, projectId }: CommentsContentProps) => {
+const CommentsItem = ({ comment, projectId }: CommentsItemProps) => {
   const { register, handleSubmit, setValue, reset } =
     useForm<CommentFormValues>()
   const [isEditing, setIsEditing] = useState(false)
+
+  const navigate = useNavigate()
+
+  const handleNavigateProfile = (userId: number) => {
+    navigate(`/profile/${userId}`)
+  }
 
   const { editCommentMutation } = useEditCommentMutation(
     Number(projectId),
@@ -66,68 +74,105 @@ const CommentsContent = ({ comment, projectId }: CommentsContentProps) => {
   }
 
   const handleError = (errors: FieldErrors<CommentFormValues>) => {
-    console.log("gg")
     console.log(errors)
   }
 
   return (
-    <Box w="100%">
-      <form onSubmit={handleSubmit(onSubmit, handleError)}>
-        <Flex
-          justifyContent="space-between"
-          w="100%">
+    <Flex
+      w="100%"
+      gap="2rem">
+      {comment.user ? (
+        <CommentsAvatar
+          onClick={() => {
+            handleNavigateProfile(comment.user.id)
+          }}
+          src={comment.user.profileImageUrl}
+        />
+      ) : (
+        <CommentsAvatar src="" />
+      )}
+
+      <Box w="100%">
+        <form onSubmit={handleSubmit(onSubmit, handleError)}>
           <Flex
-            direction="column"
-            gap="1rem"
-            flex="9.5">
+            justifyContent="space-between"
+            w="100%">
             <Flex
+              direction="column"
               gap="1rem"
-              align="center">
-              {comment.user ? (
-                <>
+              flex="9.5">
+              <Flex
+                gap="1rem"
+                align="center">
+                {comment.user ? (
+                  <>
+                    <Text
+                      fontFamily="SCDream_Bold"
+                      fontSize="xl">
+                      {comment.user.nickname}
+                    </Text>
+                  </>
+                ) : (
                   <Text
                     fontFamily="SCDream_Bold"
                     fontSize="xl">
-                    {comment.user.nickname}
+                    익명
                   </Text>
-                </>
-              ) : (
+                )}
                 <Text
-                  fontFamily="SCDream_Bold"
-                  fontSize="xl">
-                  익명
+                  color="grey.500"
+                  fontSize="md">
+                  {dateToTimeago(comment.createdAt)}
                 </Text>
-              )}
-              <Text
-                color="grey.500"
-                fontSize="md">
-                {dateToTimeago(comment.createdAt)}
-              </Text>
+              </Flex>
+              <CommentsInputOrText
+                register={register("content")}
+                isEditing={isEditing}
+                content={comment.content}
+              />
             </Flex>
-            <CommentsInputOrText
-              register={register("content")}
-              isEditing={isEditing}
-              content={comment.content}
-            />
+            <Flex
+              gap="1rem"
+              flex="0.5"
+              height="fit-content">
+              <CommentsButton
+                isOwner={comment.isOwner}
+                commentId={comment.id}
+                isEditing={isEditing}
+                handleDelete={handleDelete}
+                handleOnEdit={handleOnEdit}
+                handleOffEdit={handleOffEdit}
+              />
+            </Flex>
           </Flex>
-          <Flex
-            gap="1rem"
-            flex="0.5"
-            height="fit-content">
-            <CommentsButton
-              isOwner={comment.isOwner}
-              commentId={comment.id}
-              isEditing={isEditing}
-              handleDelete={handleDelete}
-              handleOnEdit={handleOnEdit}
-              handleOffEdit={handleOffEdit}
-            />
-          </Flex>
-        </Flex>
-      </form>
-    </Box>
+        </form>
+      </Box>
+    </Flex>
   )
 }
 
 // eslint-disable-next-line react-refresh/only-export-components
-export default withProjectId(CommentsContent)
+export default withProjectId(CommentsItem)
+
+/* <Flex
+justifyContent="space-between"
+align="center"
+key={comment.id}>
+<Flex
+  gap="2rem"
+  w="100%"
+  align="flex-start">
+  {comment.user ? (
+    <CommentsAvatar
+      onClick={() => {
+        handleNavigateProfile(comment.user.id)
+      }}
+      src={comment.user.profileImageUrl}
+    />
+  ) : (
+    <CommentsAvatar src="" />
+  )}
+
+  <CommentsContent comment={comment} />
+</Flex>
+</Flex> */
